requirements.txt
requests==2.31.0
environs==9.5.0
python_telegram_bot==13.15
flake8>=5.0.4
flake8-blind-except>=0.2.1
flake8-builtins>=2.1.0
flake8-cognitive-complexity>=0.1.0
flake8-commas>=2.1.0
flake8-comprehensions>=3.11.1
flake8-eradicate>=1.4.0
flake8-expression-complexity>=0.0.11
flake8-logging-format>=0.9.0
flake8-variables-names>=0.0.5